ai_service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Arthur Tsang",email = "arthur@youramaryllis.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "langgraph (>=0.5.3,<0.6.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "transformers (>=4.53.2,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "accelerate (>=1.8.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "protobuf (>=6.31.1,<7.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "numpy (==1.26.4)",
    "pandas (>=2.3.1,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "scikit-learn (>=1.7.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
